+++ CHANGELOG.md
@@ -4,6 +4,11 @@ See [here](http://keepachangelog.com/) for the change log format.
 
 ## [unreleased]
 
+### BREAKING
+
+- You must now use `requires` instead of the vector notation for required paths.
+- "Forking" paths must be indicated by `through-each` or `each-of`, not vectors.
+
 ### Added
 
 - a keyword `:a` in a condensed type description is equivalent to `[:a]`.
@@ -12,8 +17,10 @@ See [here](http://keepachangelog.com/) for the change log format.
 - Predicates can be applied to the whole structure.
 - `origin` and `description` let you see the definition of a type.
 - predicate sequences can include types (convenience)
+- You can refer to the paths within a type or map with `paths-of`.
+- You can use integers in paths `{[:a 1 :b] even?}`
 
 ### Changed
 
 - Namespaces reshuffled so all the pieces needed for writing custom predicates
-  are in one subdirectory, `pred-writing`.
+  are in one subdirectory, `assist`.